BloggingPlatform/BloggingAPI/serializers.py
from rest_framework import serializers
from baseModel.models import BlogPost

class BlogPostSerializer(serializers.ModelSerializer):
    class Meta:
        model = BlogPost
        fields = '__all__'