requirements.txt
Django==5.1.5
django_environ==0.12.0
djangorestframework==3.16.0
psycopg==3.1.18
psycopg2-binary==2.9.9